Updated version number, build settings and removed annoying tooltips.

+++ SS_PrefsController.m
@@ -10,7 +10,7 @@
 // ************************************************
 
 
-+ (int)version
++ (NSInteger)version
 {
     return 1; // 28th June 2003
 }
